requirements.txt
asn1crypto==1.4.0
astunparse==1.6.3
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall==0.2.0
beautifulsoup4==4.9.3
blis @ file:///C:/ci/cython-blis_1613319550054/work
breadability==0.1.20
brotlipy==0.7.0
bs4==0.0.1
cached-property==1.5.2
catalogue==1.0.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1613247269562/work
chardet @ file:///C:/ci/chardet_1607706912142/work
click @ file:///tmp/build/80754af9/click_1621604852318/work
coincurve==15.0.0
colorama==0.4.4
cryptography @ file:///C:/ci/cryptography_1616769514516/work
cymem @ file:///C:/ci/cymem_1613319479701/work
Cython==0.29.21
cytoolz==0.11.0
dataclasses==0.8
decorator==5.0.9
docopt==0.6.2
eciespy==0.3.10
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
eth-hash==0.3.1
eth-keys==0.3.3
eth-typing==2.2.2
eth-utils==1.10.0
flatbuffers==1.12
gast==0.4.0
gensim==3.8.1
grpcio==1.34.1
h5py==3.1.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877476772/work
ipykernel==5.5.3
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
jsonschema==3.0.2
jupyter-client==6.1.12
jupyter-core==4.7.1
keras-nightly==2.5.0.dev2021032900
lxml==4.6.3
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
murmurhash @ file:///C:/ci/murmurhash_1607456164222/work
nest-asyncio==1.5.1
nltk @ file:///tmp/build/80754af9/nltk_1621347441292/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1603480701039/work
pandas @ file:///C:/ci/pandas_1608056614942/work
parso==0.8.2
pickleshare==0.7.5
plac @ file:///C:/ci/plac_1594260068208/work
preshed @ file:///C:/ci/preshed_1621861011437/work
prompt-toolkit==3.0.3
pycountry==20.7.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pycryptodome==3.10.1
Pygments==2.8.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
PyQt5==5.12
PyQt5-sip==4.19.19
PyQtWebEngine==5.12
pyrsistent @ file:///C:/ci/pyrsistent_1600141799440/work
PySocks @ file:///C:/ci/pysocks_1605305839978/work
python-dateutil==2.8.1
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
pywin32==300
pyzmq==22.0.3
regex @ file:///C:/ci/regex_1617569913547/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
scipy==1.5.4
six==1.15.0
smart-open==5.1.0
soupsieve==2.2.1
spacy @ file:///C:/ci/spacy_1608303000480/work
srsly @ file:///C:/ci/srsly_1607548189686/work
sumy==0.8.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
thinc @ file:///C:/ci/thinc_1607710275191/work
toolz==0.11.1
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1615925068909/work
traitlets==4.3.3
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
wasabi @ file:///tmp/build/80754af9/wasabi_1612219178408/work
wcwidth==0.2.5
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306197271/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
